home2, board UI
