+++ frontend/src/App.js
@@ -1,13 +1,14 @@
 import "./App.css";
-import Shelter from "./components/pages/shelter/Shelter";
-import Home from "./components/pages/home/Home";
+import Home2 from "./components/pages/home2/Home2";
+import Board from "./components/pages/board/Board";
 import Shelter from './components/pages/shelter/Shelter';
 import SignUp from './components/pages/signup/SignUp';
 
 function App() {
   return (
     <>
-      <Home />
+      <Home2 />
+      <Board />
       <Shelter />
       <br />
       <SignUp />
