+++ frontend/src/components/pages/board/Board.scss
@@ -1,0 +1,21 @@
+.bgBoard {
+  background-color: aliceblue;
+}
+
+.nav {
+  width: 600px;
+  display: flex;
+  padding: 10px;
+  margin: 10px;
+  background-color: pink;
+  justify-content: space-around;
+}
+
+.category {
+  width: 600px;
+  display: flex;
+  padding: 10px;
+  margin: 10px;
+  justify-content: space-evenly;
+}
+
